Conflicts:
	view/php/theme_init.php
	view/tpl/photo_view.tpl

+++ include/hubloc.php
@@ -33,7 +33,7 @@ function prune_hub_reinstalls() {
 				// allow some slop period, say 3 days - just in case this is a glitch or transient occurrence
 				// Then remove any hublocs pointing to the oldest entry.
 
-				if($d1 < $d2) {
+				if(($d1 < $d2) && ($x[0]['hubloc_sitekey'])) {
 					logger('prune_hub_reinstalls: removing dead hublocs at ' . $rr['site_url']);
 					$y = q("delete from hubloc where hubloc_sitekey = '%s'",
 						dbesc($x[0]['hubloc_sitekey'])
@@ -96,7 +96,7 @@ function remove_obsolete_hublocs() {
 			? intval(get_config('system','delivery_interval')) : 2 );
 
 	foreach($r as $rr) {
-		q("update hubloc set hubloc_flags = (hubloc_flags | %d) where hubloc_id = %d limit 1",
+		q("update hubloc set hubloc_flags = (hubloc_flags | %d) where hubloc_id = %d",
 			intval(HUBLOC_FLAGS_DELETED),
 			intval($rr['hubloc_id'])
 		);
@@ -113,4 +113,149 @@ function remove_obsolete_hublocs() {
 }
 
 
+// This actually changes other structures to match the given (presumably current) hubloc primary selection
+
+function hubloc_change_primary($hubloc) {
+
+	if(! is_array($hubloc)) {
+		logger('no hubloc');
+		return false;
+	}
+	if(! ($hubloc['hubloc_flags'] & HUBLOC_FLAGS_PRIMARY)) {
+		logger('not primary: ' . $hubloc['hubloc_url']);
+		return false;
+	}
+
+	logger('setting primary: ' . $hubloc['hubloc_url']);
+
+	// See if there's a local channel
+
+	$r = q("select channel_id, channel_primary from channel where channel_hash = '%s' limit 1",
+		dbesc($hubloc['hubloc_hash'])
+	);
+	if(($r) && (! $r[0]['channel_primary'])) {
+		q("update channel set channel_primary = 1 where channel_id = %d",
+			intval($r[0]['channel_id'])
+		);
+	}
+
+	// do we even have an xchan for this hubloc and if so is it already set as primary?
+
+	$r = q("select * from xchan where xchan_hash = '%s' limit 1",
+		dbesc($hubloc['hubloc_hash'])
+	);
+	if(! $r) {
+		logger('xchan not found');		
+		return false;
+	}
+	if($r[0]['xchan_addr'] === $hubloc['hubloc_addr']) {
+		logger('xchan already changed');
+		return false;
+	}
+
+	$url = $hubloc['hubloc_url'];
+	$lwebbie = substr($hubloc['hubloc_addr'],0,strpos($hubloc['hubloc_addr'],'@'));
+
+	$r = q("update xchan set xchan_addr = '%s', xchan_url = '%s', xchan_follow = '%s', xchan_connurl = '%s' where xchan_hash = '%s'",
+		dbesc($hubloc['hubloc_addr']),
+		dbesc($url . '/channel/' . $lwebbie),
+		dbesc($url . '/follow?f=&url=%s'),
+		dbesc($url . '/poco/' . $lwebbie),
+		dbesc($hubloc['hubloc_hash'])
+	);
+	if(! $r)
+		logger('xchan_update failed.');
+
+	logger('primary hubloc changed.' . print_r($hubloc,true),LOGGER_DEBUG);
+	return true;
+
+}
 	
+
+function xchan_store($arr) {
+
+	if(! $arr['hash'])
+		$arr['hash'] = $arr['guid'];
+	if(! $arr['hash'])
+		return false;
+
+	$r = q("select * from xchan where xchan_hash = '%s' limit 1",
+		dbesc($arr['hash'])
+	);
+	if($r)
+		return true;
+
+	if(! $arr['network'])
+		$arr['network'] = 'unknown';
+	if(! $arr['name'])
+		$arr['name'] = 'unknown';
+	if(! $arr['url'])
+		$arr['url'] = z_root();
+	if(! $arr['photo'])
+		$arr['photo'] = get_default_profile_photo();
+
+	$r = q("insert into xchan ( xchan_hash, xchan_guid, xchan_guid_sig, xchan_pubkey, xchan_addr, xchan_url, xchan_connurl, xchan_follow, xchan_connpage, xchan_name, xchan_network, xchan_instance_url, xchan_flags, xchan_name_date ) values ('%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s','%s','%s','%s',%d,'%s') ",
+		dbesc($arr['hash']),
+		dbesc($arr['guid']),
+		dbesc($arr['guid_sig']),
+		dbesc($arr['pubkey']),
+		dbesc($arr['address']),
+		dbesc($arr['url']),
+		dbesc($arr['connurl']),
+		dbesc($arr['follow']),
+		dbesc($arr['connpage']),
+		dbesc($arr['name']),
+		dbesc($arr['network']),
+		dbesc($arr['instance_url']),
+		intval($arr['flags']),
+		dbesc(datetime_convert())
+	);
+	if(! $r)
+		return $r;
+
+	$photos = import_profile_photo($arr['photo'],$arr['hash']);
+	$r = q("update xchan set xchan_photo_date = '%s', xchan_photo_l = '%s', xchan_photo_m = '%s', xchan_photo_s = '%s', xchan_photo_mimetype = '%s' where xchan_hash = '%s'",
+		dbesc(datetime_convert()),
+		dbesc($photos[0]),
+		dbesc($photos[1]),
+		dbesc($photos[2]),
+		dbesc($photos[3]),
+		dbesc($arr['hash'])
+	);
+	return $r;
+
+}
+
+
+function xchan_fetch($arr) {
+
+	$key = '';
+	if($arr['hash']) {
+		$key = 'xchan_hash';
+		$v = $arr['hash'];
+	}
+	elseif($arr['guid']) {
+		$key = 'xchan_guid';
+		$v = $arr['guid'];
+	}
+	elseif($arr['address']) {
+		$key = 'xchan_addr';
+		$v = $arr['address'];
+	}
+
+	if(! $key)
+		return false;
+
+	$r = q("select * from xchan where $key = '$v'");
+	if(! $r)
+		return false;
+
+	$ret = array();
+	foreach($r as $k => $v) {
+		if($k === 'xchan_addr')
+			$ret['address'] = $v;
+		else
+			$ret[str_replace('xchan_','',$k)] = $v;
+	}
+	return $ret;
+}
